requirements_deep_seek.txt
langchain-huggingface
faiss-cpu
huggingface-hub
sentence-transformers
langchain-community
ollama
